Code refactoring

+++ GraphQL06/microservices/review-service/src/main/java/com/antonio/microservices/core/review/controller/ReviewControllerImpl.java
@@ -1,22 +1,23 @@
-package com.antonio.microservices.core.review.services;
-
-import java.util.List;
+package com.antonio.microservices.core.review.controller;
 
 import com.antonio.microservices.core.review.mapper.ReviewMapper;
+import com.antonio.microservices.core.review.services.ReviewServiceImpl;
+import com.antonio.microservices.core.review.web.exceptions.NotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.stereotype.Controller;
 import com.antonio.microservices.core.review.web.dto.Review;
 import com.antonio.microservices.core.review.web.exceptions.InvalidInputException;
 import com.antonio.microservices.core.review.persistence.ReviewEntity;
 import com.antonio.microservices.core.review.persistence.ReviewRepository;
 import com.antonio.microservices.core.review.web.errors.ServiceUtil;
 
-@RestController
-public class ReviewServiceImpl implements ReviewService {
+import java.util.List;
 
+@Controller
+public class ReviewControllerImpl implements ReviewController {
     private static final Logger LOG = LoggerFactory.getLogger(ReviewServiceImpl.class);
 
     private final ReviewRepository repository;
@@ -26,23 +27,23 @@ public class ReviewServiceImpl implements ReviewService {
     private final ServiceUtil serviceUtil;
 
     @Autowired
-    public ReviewServiceImpl(ReviewRepository repository, ReviewMapper mapper, ServiceUtil serviceUtil) {
+    public ReviewControllerImpl(ReviewRepository repository, ReviewMapper mapper, ServiceUtil serviceUtil) {
         this.repository = repository;
         this.mapper = mapper;
         this.serviceUtil = serviceUtil;
     }
 
     @Override
-    public Review createReview(Review body) {
+    public Review createReviews(Review input) {
         try {
-            ReviewEntity entity = mapper.apiToEntity(body);
+            ReviewEntity entity = mapper.apiToEntity(input);
             ReviewEntity newEntity = repository.save(entity);
 
-            LOG.debug("createReview: created a review entity: {}/{}", body.getProductId(), body.getReviewId());
+            LOG.debug("createReview: created a review entity: {}/{}", input.getProductId(), input.getReviewId());
             return mapper.entityToApi(newEntity);
 
         } catch (DataIntegrityViolationException dive) {
-            throw new InvalidInputException("Duplicate key, Product Id: " + body.getProductId() + ", Review Id:" + body.getReviewId());
+            throw new InvalidInputException("Duplicate key, Product Id: " + input.getProductId() + ", Review Id:" + input.getReviewId());
         }
     }
 
@@ -56,14 +57,23 @@ public class ReviewServiceImpl implements ReviewService {
         List<ReviewEntity> entityList = repository.findByProductId(productId);
         List<Review> list = mapper.entityListToApiList(entityList);
 
+        if (list.size() == 0) {
+            LOG.debug("getReviews: no reviews found for productId: {}", productId);
+            throw new NotFoundException("No reviews found for productId: " + productId);
+        }
         LOG.debug("getReviews: response size: {}", list.size());
 
         return list;
     }
 
     @Override
-    public void deleteReviews(int productId) {
+    public Boolean deleteReviews(int productId) {
         LOG.debug("deleteReviews: tries to delete reviews for the product with productId: {}", productId);
+        if (repository.findByProductId(productId).isEmpty()) {
+            throw new NotFoundException("No reviews found for productId: " + productId);
+        }
         repository.deleteAll(repository.findByProductId(productId));
+
+        return Boolean.TRUE;
     }
 }
